backend/backend_services/user_managment/app/requirements.txt
django
django-oauth-toolkit
psycopg2
djangorestframework
djangorestframework-simplejwt
pyjwt
gunicorn
